Thanks to Frans Meulenbroeks
http://groups.google.com/group/opkg-devel/browse_thread/thread/23c3557277de0f2e

If a file name in a tar archive is exactly 100 bytes long the name
field is completely filled and there is no terminating null byte;
so extraction of the file will yield a name that is extended with the
mode (e.g. 000644).

The attached patch cures it although there might be better solutions.
The problem is also in busybox tar and reported there too.

Frans.

git-svn-id: https://opkg.googlecode.com/svn/trunk@201 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358