Fix: RRECOMMENDS and RDEPENDS being basically the same
While removing a package with opkg, the process shouldn't be blocked if
another package RECOMMENDS the package wanted to be removed. This is
because, while generating the dependencies, opkg adds dependencies to
depended_upon_by even if dependency's type is RECOMMEND. The fix is to
skip dependencies of type RECOMMEND while constructing depended_upon_by.

Bug info:
https://bugzilla.yoctoproject.org/show_bug.cgi?id=2431

Signed-off-by: Andrei Gherzan <andrei@gherzan.ro>